refactor: PEP 604

+++ cogniq/slack/cogniq_slack.py
@@ -151,7 +151,7 @@ class CogniqSlack:
         uvicorn_server = uvicorn.Server(uvicorn_config)
         await uvicorn_server.serve()
 
-    async def chat_update(self, *, channel: str, ts: int, context: Dict, text: str, retry_on_rate_limit: bool = True):
+    async def chat_update(self, *, channel: str, ts: float, context: Dict, text: str, retry_on_rate_limit: bool = True):
         """
         Updates the chat message in the given channel and thread with the given text.
         """
